Spanish translation... "upgraded"? :-).
Just FYI of what has been changed. For the noun "survey" there are two very common used words in spanish, "encuesta" and "cuestionario". Everybody understand both almost as the same thing, but "cuestionario" much like "questionnaire", has an indirect meaning of "list of questions" and "encuesta" usually means something bigger: the development of the survey's questionnaire and any further work on the whole process of surveying, including the act of responding to the survey and its results.

So, I changed almost all the "cuestionario" appearances on behalf of "encuesta" (which also leads me to change the corresponding gender of surrounding articles and pronouns from male to female).

Hopefully, it will help to understand better some questions asked by LimeSurvey to spanish parlant survey administrators.